Tweak reflector config
I went on a demolinux VM. I wanted to install a package. Guess what? After a few seconds it timed out. I was about to insult Reflector for choosing shit mirrors. So I tried to change the --sort option. I tried installing the package again and it didn't time out. Does that mean changing this option was actually useful? Perhaps.